backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
motor==3.4.0
pymongo==4.6.3
pydantic==2.6.4
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
bcrypt==4.1.2
passlib==1.7.4
python-jose==3.3.0
cryptography==42.0.5
requests==2.31.0
openai==1.14.3
googlemaps==4.10.0
amadeus==8.1.0
aiohttp==3.9.3
typer==0.9.0
rich==13.7.1
click==8.1.7
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
anyio==4.3.0
sniffio==1.3.1
h11==0.14.0
starlette==0.37.2
typing-extensions==4.10.0
annotated-types==0.6.0
pydantic-core==2.16.3
